+++ src/patch/common/net_hook.h
@@ -30,7 +30,7 @@ void NetPatch_pushMod(const char* modId, int version);
 void NetPatch_overrideReloadMods(bool val);
 void NetPatch_overrideModList();
 
-void NetPatch_pushDLC(int data1, int data2, int data3, int data4);
+void NetPatch_pushDLC(uint32_t data1, uint16_t data2, uint16_t data3, uint64_t data4);
 void NetPatch_overrideReloadDLC(bool val);
 void NetPatch_overrideDLCList();
